feat(bars)! Remove incline and move trouble to lualine winbar
Incline was nice but added overhead.
Trouble was originally in the bottom lualine statusline, but that looked
messy when the trouble symbols were long.
So, moved those to the winbar, but looked strange on their own, so added
the filename component there.
This makes incline redundant, and so it was removed.